Small patch for COMPRESSED reading/writing & small big fix

+++ EventReconstruction/Calo/include/LCaloSignal.hh
@@ -28,10 +28,10 @@ public:
   int GetThe2ndMostSignificantUnit(const bool isHG, const double threshold = THRESHOLD) const;
   double GetSNOfMSU(const bool isHG, const double threshold = THRESHOLD) const;
   double GetSNOf2ndMSU(const bool isHG, const double threshold = THRESHOLD) const;
-  double GetCountsOfMSU(const bool isHG, const double threshold = THRESHOLD) const;
-  double GetCountsOf2ndMSU(const bool isHG, const double threshold = THRESHOLD) const;
+  int GetCountsOfMSU(const bool isHG, const double threshold = THRESHOLD) const;
+  int GetCountsOf2ndMSU(const bool isHG, const double threshold = THRESHOLD) const;
   double GetSNOfUnit(const int unit, const bool isHG) const;
-  double GetCountsOfUnit(const int unit, const bool isHG) const;
+  int GetCountsOfUnit(const int unit, const bool isHG) const;
   
 protected:
   void CreateContainers(void);
